Extended definition of Simplex to follow recommendations for constructing collections
Based on https://docs.scala-lang.org/overviews/core/custom-collections.html this commit includes new test cases that illustrate failure of preserving type for a number of inherited methods, and restructures the code to closer work as an extension of `TreeSet` with the functionality implemented in a way that automatically draws on the inherited functions from `SortedSet`.
Along the way, we found a need to represent `Simplex` not just as a pure `Int`-container, but to make the notion of `Simplex` containing arbitrary (orderable) objects available - in order to work well with the type preservation setup. So now `Simplex` is a type alias for `AbstractSimplex[Int]`, and we can construct `AbstractSimplex` containing all sorts of (implicitly) orderable types.