Moved filtrationValues from the Simplex class to the SimplexStream class
A filtration is a property of a simplicial complex, contributing to a total order on the simplices in that simplicial complex. As such, it makes more sense to make filtration values not be a field on a simplex object, but be a partial function that assigns to all valid enough simplices a filtration value.
This change also makes it much easier to deal with filtration values interacting with boundary maps: we automatically get the appropriate filtration value for any subsimplex.